+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2013-12-10
+
+* Support for downloading points as DSV
+
 ## 2013-11-25
 
 * Added ability to rename GeoJSON properties in the table view